+++ src/helpers/Db.php
@@ -1079,10 +1079,13 @@ class Db
     /**
      * Resets the memoized database connection.
      *
-     * @since 3.5.13
+     * @since 3.5.12.1
      */
     public static function reset()
     {
+        if (self::$_db) {
+            self::$_db->close();
+        }
         self::$_db = null;
     }