+++ src/migrations/Install.php
@@ -1046,10 +1046,7 @@ class Install extends Migration
 
         $applyExistingProjectConfig = false;
 
-        if (
-            $this->applyProjectConfigYaml &&
-            file_exists($configFile = Craft::$app->getPath()->getProjectConfigFilePath())
-        ) {
+        if ($this->applyProjectConfigYaml && $projectConfig->getDoesYamlExist()) {
             try {
                 $expectedSchemaVersion = (string)$projectConfig->get(ProjectConfig::CONFIG_SCHEMA_VERSION_KEY, true);
                 $craftSchemaVersion = (string)Craft::$app->schemaVersion;
@@ -1060,7 +1057,7 @@ class Install extends Migration
                 }
 
                 // Make sure at least sites are processed
-                ProjectConfigHelper::ensureAllSitesProcessed();
+                ProjectConfigHelper::ensureAllSitesProcessed(true);
 
                 $this->_installPlugins();
                 $applyExistingProjectConfig = true;