+++ src/services/Elements.php
@@ -468,13 +468,48 @@ class Elements extends Component
      */
     public function getElementById(int $elementId, string $elementType = null, $siteId = null, array $criteria = [])
     {
+        return $this->_elementById('id', $elementId, $elementType, $siteId, $criteria);
+    }
+
+    /**
+     * Returns an element by its UID.
+     *
+     * If no element type is provided, the method will first have to run a DB query to determine what type of element
+     * the $uid is, so you should definitely pass it if it’s known.
+     * The element’s status will not be a factor when using this method.
+     *
+     * @param string $uid The element’s UID.
+     * @param string|null $elementType The element class.
+     * @param int|int[]|string|null $siteId The site(s) to fetch the element in.
+     * Defaults to the current site.
+     * @param array $criteria
+     * @return ElementInterface|null The matching element, or `null`.
+     * @since 3.5.13
+     */
+    public function getElementByUid(string $uid, string $elementType = null, $siteId = null, array $criteria = [])
+    {
+        return $this->_elementById('uid', $uid, $elementType, $siteId, $criteria);
+    }
+
+    /**
+     * Returns an element by its ID or UID.
+     *
+     * @param string $property Either `id` or `uid`
+     * @param int|string $elementId The element’s ID/UID
+     * @param string|null $elementType The element class.
+     * @param int|int[]|string|null $siteId The site(s) to fetch the element in.
+     * Defaults to the current site.
+     * @param array $criteria
+     * @return ElementInterface|null The matching element, or `null`.
+     */
+    private function _elementById(string $property, $elementId, string $elementType = null, $siteId = null, array $criteria = [])
+    {
         if (!$elementId) {
             return null;
         }
 
         if ($elementType === null) {
-            /** @noinspection CallableParameterUseCaseInTypeContextInspection */
-            $elementType = $this->getElementTypeById($elementId);
+            $elementType = $this->_elementTypeById($property, $elementId);
 
             if ($elementType === null) {
                 return null;
@@ -486,7 +521,7 @@ class Elements extends Component
         }
 
         $query = $this->createElementQuery($elementType);
-        $query->id = $elementId;
+        $query->$property = $elementId;
         $query->siteId = $siteId;
         $query->anyStatus();
 
@@ -495,7 +530,7 @@ class Elements extends Component
             $data = (new Query())
                 ->select(['draftId', 'revisionId'])
                 ->from([Table::ELEMENTS])
-                ->where(['id' => $elementId])
+                ->where([$property => $elementId])
                 ->one();
         } catch (DbException $e) {
             // Not on schema 3.2.6+ yet
@@ -587,10 +622,34 @@ class Elements extends Component
      */
     public function getElementTypeById(int $elementId)
     {
+        return $this->_elementTypeById('id', $elementId);
+    }
+
+    /**
+     * Returns the class of an element with a given UID.
+     *
+     * @param string $uid The element’s UID
+     * @return string|null The element’s class, or null if it could not be found
+     * @since 3.5.13
+     */
+    public function getElementTypeByUid(string $uid)
+    {
+        return $this->_elementTypeById('uid', $uid);
+    }
+
+    /**
+     * Returns the class of an element with a given ID/UID.
+     *
+     * @param string $property Either `id` or `uid`
+     * @param int|string $elementId The element’s ID/UID
+     * @return string|null The element’s class, or null if it could not be found
+     */
+    private function _elementTypeById(string $property, $elementId)
+    {
         $class = (new Query())
             ->select(['type'])
             ->from([Table::ELEMENTS])
-            ->where(['id' => $elementId])
+            ->where([$property => $elementId])
             ->scalar();
 
         return $class !== false ? $class : null;