+++ src/utilities/SystemReport.php
@@ -71,10 +71,23 @@ class SystemReport extends Utility
             }
         }
 
+        $aliases = [];
+        foreach (Craft::$aliases as $alias => $value) {
+            if (is_array($value)) {
+                foreach ($value as $a => $v) {
+                    $aliases[$a] = $v;
+                }
+            } else {
+                $aliases[$alias] = $value;
+            }
+        }
+        ksort($aliases);
+
         return Craft::$app->getView()->renderTemplate('_components/utilities/SystemReport', [
             'appInfo' => self::_appInfo(),
             'plugins' => Craft::$app->getPlugins()->getAllPlugins(),
             'modules' => $modules,
+            'aliases' => $aliases,
             'requirements' => self::_requirementResults(),
         ]);
     }